+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 * Added `itertools.js` to gulpfile build files in order to be bundled
+* `getPrices` which consumes `/api/config/prices` for getting prices for several configs in a single batch call
 
 ### Changed